implement Announcement function(CR)

+++ src/main/java/com/limbae/pfy/domain/AnnouncementVO.java
@@ -18,7 +18,7 @@ public class AnnouncementVO {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int idx;
+    Long idx;
 
     String title;
